state_manager/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "state-manager"
version = "0.1.0"
description = "Exosphere State Manager"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "apscheduler>=3.11.0",
    "beanie>=2.0.0",
    "croniter>=6.0.0",
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "json-schema-to-pydantic>=0.4.1",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

# Tell Poetry which package to install
packages = [
    { include = "state_manager" }
]

[tool.poetry]
# optional metadata for Poetry
authors = ["Your Name <youremail@example.com>"]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "asgi-lifespan>=2.1.0",
]
